[authorized_key] adds state=replace_all option
The authorized_key module enable to add (state=present) or remove
(state=absent) one or more keys to the authorized_keys file.
The fact that the module supports multiple keys is not documented
but works if you separate your keys with a newline "\n" character.

This commit adds a third state=replace_all which ensures that only
the added keys will be present in the authorized_keys and removes
all the others.